requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
diskcache==5.6.3
distro==1.9.0
docstring_parser==0.17.0
frozenlist==1.7.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
instructor==1.10.0
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-core==0.3.74
langchain-openai==0.3.30
langchain-text-splitters==0.3.9
langsmith==0.4.14
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.6.4
openai==1.99.9
orjson==3.11.2
packaging==25.0
pluggy==1.6.0
propcache==0.3.2
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pytest==8.4.1
python-dotenv==1.1.1
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.1.0
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.43
tenacity==9.1.2
tiktoken==0.11.0
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
yarl==1.20.1
zstandard==0.23.0
